#9: renamed module pathnames to connection_pathnames. refactoring code.
