+++ lib/el_finder/connector/local_file_system.rb
@@ -50,7 +50,7 @@ module ElFinder
         raise(ArgumentError, "Mime Handler is invalid") unless mime_handler.respond_to?(:for)
         raise(ArgumentError, "Image Handler is invalid") unless image_handler.nil? || ([:size, :resize, :thumbnail].all?{|m| image_handler.respond_to?(m)})
 
-        @root = ElFinder::Pathnames::FileSystem.new(options[:root])
+        @root = ElFinder::ConnectionPathnames::FileSystemPathname.new(options[:root])
 
         @options[:volume_id] = "#{DRIVER_ID}#{@options[:index]}" unless @options[:index] == 0
 
@@ -620,7 +620,7 @@ module ElFinder
 
       #
       def specific_perm_for(pathname, perm)
-        pathname = pathname.path if pathname.is_a?(ElFinder::Pathnames::FileSystem)
+        pathname = pathname.path if pathname.is_a?(ElFinder::ConnectionPathnames::FileSystemPathname)
         matches = @options[:perms].select{ |k,v| pathname.to_s.send((k.is_a?(String) ? :== : :match), k) }
         if perm == :hidden
           matches.one?{|e| e.last[perm] }
